only use node:fs in one file
